De-duplicated the version override in the actual project and in buildSrc
